Revert moving `AnalyticsStore` to `api` module (twice).
Use `AnalyticsService` in `DefaultCallWidgetSettingsProvider` instead.
